added sync directive
