+++ .bramble/pipelines/hello.yaml
@@ -7,5 +7,6 @@ metadata:
 spec:
   tasks:
   - name: hello
-    image: alpine
-    command: ["echo", "hello", "world"]
+    spec:
+      image: alpine
+      command: ["echo", "hello", "world"]
